refactor(ascii-term): split terminal.rs and audio.rs by responsibility
